dbsp_adapters: Move `InputCommandReceiver` to feldera-adapterlib.
An upcoming commit will use it from feldera-datagen.
